More work on hinting.
The problem is knowing when to drag the value over bb end in reg
and when to flush it to mem. Flashing to mem should happen on the
last instruction, before that the value should be dragged through.
To implement we need to find the last instruction that uses the
selected variable and allow liveness to take effect (mark as dead)